\name{package:CNtu}
\docType{package}
\alias{CNtu}
\title{
	Copy Number Tree Utilities
}
\description{
Utility functions for copy number data
}
\details{
	\tabular{ll}{
		Package: \tab CNTU\cr
		Type: \tab Package\cr
		Version: \tab 0.0-2\cr
		Date: \tab 2018-02-05\cr
		License: \tab GPL (>= 2)\cr
		LazyLoad: \tab yes\cr
	}
}
\author{
	\tabular{ll}{
		author(s): \tab David Brown\cr
				   \tab brownd7@mskcc.org\cr
				   \tab \cr
		maintainer(s): \tab David Brown\cr
					   \tab brownd7@mskcc.org\cr
	}
}
\keyword{package}
