\name{package:CNtu}
\docType{package}
\alias{CNtu}
\title{
	Copy Number Tree Utilities
}
\description{
Utility functions for copy number data
}
\details{
	\tabular{ll}{
		Package: \tab CNTU\cr
		Type: \tab Package\cr
		Version: \tab 0.0-1\cr
		Date: \tab 2014-11-14\cr
		License: \tab GPL (>= 2)\cr
		LazyLoad: \tab yes\cr
	}
}
\author{
	\tabular{ll}{
		author(s): \tab David Brown\cr
				   \tab david.brown@bordet.be\cr
				   \tab \url{http://homepages.ulb.ac.be/~dbrown/}\cr\cr
				   \tab \cr
		maintainer(s): \tab David Brown\cr
					   \tab david.brown@bordet.be\cr
					   \tab \url{http://homepages.ulb.ac.be/~dbrown/}\cr\cr
	}
}
\keyword{package}
